refactor(modules) rename model to schema
