+++ src/modules/customers/customer.schema.ts
